refactoring of ConfigurationBuilder
